+++ src/indexHelpers.ts
@@ -31,5 +31,6 @@ export { default as WUPScrolled } from "./helpers/scrolled";
 export { default as scrollIntoView } from "./helpers/scrollIntoView";
 export { stringPrettify, stringLowerCount, stringUpperCount } from "./helpers/string";
 
-export { default as zip } from "./helpers/zip";
-export { default as exportToExcel } from "./helpers/exportToExcel";
+export { default as zip } from "./helpers/files/zip";
+export { default as exportToExcel } from "./helpers/files/exportToExcel";
+export { default as saveAsFile } from "./helpers/files/saveAsFile";